+++ src/lib/components/select-contract/ContractListDetail.tsx
@@ -55,7 +55,8 @@ const ContractListContent = ({
           <EmptyState
             imageVariant="not-found"
             message="No matching contracts found.
-  Make sure you are searching with a contract address, name, or description."
+          Make sure you are searching with a contract address, name, or description."
+            withBorder
           />
         )
       }